Implement interrupt ignoring
In the process of trying to get interrupt ignoring to work I also found
that interrupting would sometimes cause the behavior described in
thinkOS chapter 4 about cached writes not being written when the
program crashed. I wrote another handler that closed the files before
exiting, which meant adding the files pointer array and counter to the
globals, allocating space for it with malloc, and refactoring the code
that appends to the array to incrementally update the counter.